requirements.txt
fastapi
uvicorn
requests
pyjwt
streamlit
hdbcli